+++ project/apps/bff/src/app/app.http
@@ -23,7 +23,7 @@ POST http://localhost:3002/api/users/login HTTP/1.1
 Content-Type: application/json
 
 {
-  "email": "client@gmail.local",
+  "email": "executor@gmail.local",
   "password": "123456"
 }
 
@@ -108,7 +108,7 @@ Content-Type: application/json
 ###
 
 # Change task-status
-PATCH http://localhost:3002/api/tasks/status/10 HTTP/1.1
+PATCH http://localhost:3002/api/tasks/status/8 HTTP/1.1
 Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2NGYzMWM5ZDRkNzcyMWVhYTE1ZWZhNjAiLCJlbWFpbCI6ImNsaWVudEBnbWFpbC5sb2NhbCIsInJvbGUiOiJDbGllbnQiLCJsYXN0bmFtZSI6IlNtaXRoIiwiZmlyc3RuYW1lIjoiS2VrcyIsImlhdCI6MTY5MzY1NDE5MCwiZXhwIjoxNjkzNzQwNTkwfQ.cUXI_f8C0yKuOdb9lzbzG7TKrh39v5iETa01IGgO-PA
 Content-Type: application/json
 
@@ -120,11 +120,11 @@ Content-Type: application/json
 
 ###
 # Add new response
-POST http://localhost:3002/api/tasks/response/10 HTTP/1.1
+POST http://localhost:3002/api/tasks/response/8 HTTP/1.1
 Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2NGYzNGRlMWZjZThlNTZhZTEwYTYyMGIiLCJlbWFpbCI6ImV4ZWNAZ21haWwubG9jYWwiLCJyb2xlIjoiRXhlY3V0b3IiLCJsYXN0bmFtZSI6IlNtaXRoIiwiZmlyc3RuYW1lIjoiS2VrcyIsImlhdCI6MTY5MzY2NjgwNCwiZXhwIjoxNjkzNzUzMjA0fQ.F3UJthQPsi10R-n3W63cYG-28aJLcSsW3XDZMSueFV0
 ###
 # Delete task by Id
-POST http://localhost:3002/api/tasks/delete/5 HTTP/1.1
+POST http://localhost:3002/api/tasks/delete/6 HTTP/1.1
 Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2NGYzMWM5ZDRkNzcyMWVhYTE1ZWZhNjAiLCJlbWFpbCI6ImNsaWVudEBnbWFpbC5sb2NhbCIsInJvbGUiOiJDbGllbnQiLCJsYXN0bmFtZSI6IlNtaXRoIiwiZmlyc3RuYW1lIjoiS2VrcyIsImlhdCI6MTY5MzY1NDE5MCwiZXhwIjoxNjkzNzQwNTkwfQ.cUXI_f8C0yKuOdb9lzbzG7TKrh39v5iETa01IGgO-PA
 ####
 ###---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
@@ -153,7 +153,7 @@ GET http://localhost:3002/api/comments?limit=3&page=2 HTTP/1.1
 ###---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 
 # Create a new review
-POST http://localhost:3002/api/reviews/create/10 HTTP/1.1
+POST http://localhost:3002/api/reviews/create/7 HTTP/1.1
 Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2NGYzMWM5ZDRkNzcyMWVhYTE1ZWZhNjAiLCJlbWFpbCI6ImNsaWVudEBnbWFpbC5sb2NhbCIsInJvbGUiOiJDbGllbnQiLCJsYXN0bmFtZSI6IlNtaXRoIiwiZmlyc3RuYW1lIjoiS2VrcyIsImlhdCI6MTY5MzY1NDE5MCwiZXhwIjoxNjkzNzQwNTkwfQ.cUXI_f8C0yKuOdb9lzbzG7TKrh39v5iETa01IGgO-PA
 Content-Type: application/json
