+++ project/apps/bff/src/app/categories.controller.ts
@@ -1,10 +1,10 @@
-import { Body, Controller, Get, Post, UseFilters, UseGuards, UseInterceptors } from "@nestjs/common";
+import { Body, Controller, Get, Post, UseFilters, UseGuards} from "@nestjs/common";
 import { AxiosExceptionFilter } from "./filters/axios-exception.filter";
 import { CheckAuthGuard } from "./guards/check-auth.guard";
-import { ClientInterceptor } from "./interseptors/client.intrceptor";
 import { CreateCategoryDto } from "./dto/create-category.dto";
 import { ApplicationServiceURL } from "./app.config";
 import { HttpService } from "@nestjs/axios";
+import { ClientGuard } from "./guards/client.guard";
 
 @Controller('categories')
 @UseFilters(AxiosExceptionFilter)
@@ -14,8 +14,7 @@ export class CategoriesController {
     private readonly httpService: HttpService,
   ) {}
 
-  @UseGuards(CheckAuthGuard)
-  @UseInterceptors(ClientInterceptor)
+  @UseGuards(CheckAuthGuard, ClientGuard)
   @Post('create')
   public async create(@Body() dto: CreateCategoryDto) {
     const { data } = await this.httpService.axiosRef.post(`${ApplicationServiceURL.Categories}/create`, dto);
