+++ project/apps/bff/src/app/guards/permission.guard.ts
@@ -1,22 +1,21 @@
+import { CanActivate, ExecutionContext, ForbiddenException, Injectable } from '@nestjs/common';
 import { HttpService } from '@nestjs/axios';
-import {BadRequestException, CallHandler, ExecutionContext, Injectable, NestInterceptor} from '@nestjs/common';
 import { ApplicationServiceURL } from '../app.config';
 
-
 @Injectable()
-export class PermisionInterseptor implements NestInterceptor {
+export class PermissionGuard implements CanActivate {
   constructor(
     private readonly httpService: HttpService,
   ) {}
-  
-  public async intercept(context: ExecutionContext, next: CallHandler) {
+
+  public async canActivate(context: ExecutionContext): Promise<boolean> {
     const request = context.switchToHttp().getRequest();
     const { data } = await this.httpService.axiosRef.get(`${ApplicationServiceURL.Tasks}/${request.params.id}`);
     const fromRequest = request.user.sub.toString();
     const fromData = data.creatorId.toString();
     if(fromRequest != fromData){
-      throw new BadRequestException("permission denied")
+      throw new ForbiddenException("permission denied")
     }
-    return next.handle();
+    return true;
   }
 }
