+++ project/apps/bff/src/app/guards/status.permision.guard.ts
@@ -1,16 +1,15 @@
+import { CanActivate, ExecutionContext, ForbiddenException, Injectable } from '@nestjs/common';
 import { HttpService } from '@nestjs/axios';
-import {BadRequestException, CallHandler, ExecutionContext, Injectable, NestInterceptor} from '@nestjs/common';
 import { ApplicationServiceURL } from '../app.config';
 import { TaskStatus } from '@project/shared/app-types';
 
-
 @Injectable()
-export class StatusPermisionInterseptor implements NestInterceptor {
+export class StatusPermissionGuard implements CanActivate {
   constructor(
     private readonly httpService: HttpService,
   ) {}
-  
-  public async intercept(context: ExecutionContext, next: CallHandler) {
+
+  public async canActivate(context: ExecutionContext): Promise<boolean> {
     const request = context.switchToHttp().getRequest();
     const { data } = await this.httpService.axiosRef.get(`${ApplicationServiceURL.Tasks}/${request.params.id}`);
     const fromRequest = request.user.sub.toString();
@@ -21,8 +20,8 @@ export class StatusPermisionInterseptor implements NestInterceptor {
       fromData = data.creatorId.toString();
     }
     if(fromRequest != fromData){
-      throw new BadRequestException("permission denied")
+      throw new ForbiddenException("permission denied")
     }
-    return next.handle();
+    return true;
   }
 }
