--- project/apps/bff/src/app/interseptors/creatorid.interceptor.ts
@@ -1,11 +1,0 @@
-import { CallHandler, ExecutionContext, Injectable, NestInterceptor } from '@nestjs/common';
-
-@Injectable()
-export class CreatorIdInterceptor implements NestInterceptor {
-  public intercept(context: ExecutionContext, next: CallHandler) {
-    const request = context.switchToHttp().getRequest();
-    request.body['creatorId'] = request.user.sub;
-
-    return next.handle();
-  }
-}
