+++ project/apps/bff/src/app/review.controller.ts
@@ -1,8 +1,8 @@
-import { Body, Controller, Param, Post, UseFilters, UseGuards, UseInterceptors } from "@nestjs/common";
+import { Body, Controller, Param, Post, UseFilters, UseGuards} from "@nestjs/common";
 import { AxiosExceptionFilter } from "./filters/axios-exception.filter";
 import { HttpService } from "@nestjs/axios";
 import { CheckAuthGuard } from "./guards/check-auth.guard";
-import { ReviewPermisionInterseptor } from "./interseptors/review.permision.interceptor";
+import { PermissionGuard } from "./guards/permission.guard";
 import { CreateReviewDto } from "./dto/crete-review.dto";
 import { ApplicationServiceURL } from "./app.config";
 
@@ -14,8 +14,7 @@ export class ReviewController {
     private readonly httpService: HttpService,
   ) {}
 
-  @UseGuards(CheckAuthGuard)
-  @UseInterceptors(ReviewPermisionInterseptor)
+  @UseGuards(CheckAuthGuard, PermissionGuard)
   @Post('create/:id')
   public async create(@Body() dto: CreateReviewDto, @Param('id') id: string) {
     const { data } = await this.httpService.axiosRef.post(`${ApplicationServiceURL.Reviews}/create/${id}`, dto);
